move ResponseFilterContext to app dir
