+++ apps/web/app/environments/[environmentId]/surveys/[surveyId]/ResponseFilter.tsx
@@ -7,7 +7,7 @@ import { Popover, PopoverTrigger, PopoverContent, Button, Checkbox } from "@form
 import { ChevronDown, ChevronUp, Plus } from "lucide-react";
 import { TrashIcon } from "@heroicons/react/24/solid";
 import QuestionFilterComboBox from "@/app/environments/[environmentId]/surveys/[surveyId]/QuestionFilterComboBox";
-import { useResponseFilter } from "@/lib/surveys/ResponseFilterContext";
+import { useResponseFilter } from "@/app/environments/[environmentId]/ResponseFilterContext";
 import clsx from "clsx";
 
 export type QuestionFilterOptions = {
